Rename OpenSearch REST common module
