+++ .fossa.yml
@@ -798,7 +798,7 @@ targets:
       target: ':instrumentation:opensearch:opensearch-rest-3.0:javaagent'
     - type: gradle
       path: ./
-      target: ':instrumentation:opensearch:opensearch-rest-common:javaagent'
+      target: ':instrumentation:opensearch:opensearch-rest-common-1.0:javaagent'
     - type: gradle
       path: ./
       target: ':instrumentation:opentelemetry-api:opentelemetry-api-1.0:javaagent'
