+++ instrumentation/opensearch/opensearch-java-3.0/javaagent/build.gradle.kts
@@ -21,7 +21,7 @@ dependencies {
   compileOnly("com.fasterxml.jackson.core:jackson-core")
 
   testImplementation("org.opensearch.client:opensearch-rest-client:3.0.0")
-  testImplementation(project(":instrumentation:opensearch:opensearch-rest-common:testing"))
+  testImplementation(project(":instrumentation:opensearch:opensearch-rest-common-1.0:testing"))
   testInstrumentation(project(":instrumentation:apache-httpclient:apache-httpclient-5.0:javaagent"))
 
   // AwsSdk2Transport supports awssdk version 2.26.0
