+++ instrumentation/opensearch/opensearch-rest-common-1.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/opensearch/rest/common/v1_0/RestResponseListener.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.opensearch.rest.common;
+package io.opentelemetry.javaagent.instrumentation.opensearch.rest.common.v1_0;
 
 import io.opentelemetry.context.Context;
 import io.opentelemetry.context.Scope;
