+++ settings.gradle.kts
@@ -508,8 +508,8 @@ include(":instrumentation:opencensus-shim:testing")
 include(":instrumentation:opensearch:opensearch-java-3.0:javaagent")
 include(":instrumentation:opensearch:opensearch-rest-1.0:javaagent")
 include(":instrumentation:opensearch:opensearch-rest-3.0:javaagent")
-include(":instrumentation:opensearch:opensearch-rest-common:javaagent")
-include(":instrumentation:opensearch:opensearch-rest-common:testing")
+include(":instrumentation:opensearch:opensearch-rest-common-1.0:javaagent")
+include(":instrumentation:opensearch:opensearch-rest-common-1.0:testing")
 include(":instrumentation:opentelemetry-api:opentelemetry-api-1.0:javaagent")
 include(":instrumentation:opentelemetry-api:opentelemetry-api-1.0:testing")
 include(":instrumentation:opentelemetry-api:opentelemetry-api-1.4:javaagent")
